clean up sprite class
