+++ src/scenes/Game.ts
@@ -1,6 +1,6 @@
 import Phaser from 'phaser';
-import { ISprite } from '../graphics/Sprite';
 import SpriteFactory from '../graphics/SpriteFactory';
+import { ISprite } from '../graphics/types';
 
 export default class Demo extends Phaser.Scene {
   private sprite: ISprite;
@@ -26,11 +26,14 @@ export default class Demo extends Phaser.Scene {
     });
   }
 
+  private addDelta = 0;
+
   update(time: number, delta: number): void {
     this.sprite.update(time, delta);
+    this.addDelta += delta / 1000;
 
     this.rt.clear();
-    this.fpsText.setText(`FPS: ${(1000 / delta).toFixed(3)}\nDelta: ${(delta / 1000).toFixed(4)}\nTime: ${time / 1000}`);
+    this.fpsText.setText(`FPS: ${(1000 / delta).toFixed(3)}\nDelta: ${(delta / 1000).toFixed(4)}\nTime: ${time / 1000}\nAdd Delta: ${this.addDelta}`);
     this.rt.beginDraw();
     this.sprite.draw(this.rt, new Phaser.Math.Vector2(525, 200));
     this.rt.endDraw();
